% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{getToviDataFrame}
\alias{getToviDataFrame}
\title{Conversion Function - Given datafile Id return a generic Tovi dataframe}
\usage{
getToviDataFrame(dataFileId)
}
\arguments{
\item{dataFileId}{id of a datafile}
}
\value{
DataFrame
}
\description{
Conversion Function - Given datafile Id return a generic Tovi dataframe
}
