% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{netCDFToDF}
\alias{netCDFToDF}
\title{Conversion Function - converts data in Tovi NetCDF file into generic dataframe}
\usage{
netCDFToDF(filePath)
}
\arguments{
\item{filePath}{is a character file path to a netcdf file}
}
\value{
dataframe
}
\description{
Conversion Function - converts data in Tovi NetCDF file into generic dataframe
}
