% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{netCDFToREddyProcDF}
\alias{netCDFToREddyProcDF}
\title{Conversion Function - converts data in Tovi NetCDF file into dataframe for REddyProc}
\usage{
netCDFToREddyProcDF(x)
}
\arguments{
\item{filePath}{is a character file path to a netcdf file}
}
\value{
dataframe
}
\description{
Conversion Function - converts data in Tovi NetCDF file into dataframe for REddyProc
}
