% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{toviDataToREddyProcDF}
\alias{toviDataToREddyProcDF}
\title{Conversion Function - converts data in Tovi netcdf or dataframe into REddyProc formatted dataframe}
\usage{
toviDataToREddyProcDF(x)
}
\arguments{
\item{x}{is a file path to NetCDF file or a dataframe}
}
\value{
dataframe
}
\description{
Conversion Function - converts data in Tovi netcdf or dataframe into REddyProc formatted dataframe
}
\examples{
toviDataToREddyProcDF('path/to/file.nc')
toviDataToREddyProcDF(df)
}
