% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{dfToREddyProcClass}
\alias{dfToREddyProcClass}
\title{Conversion Function - converts data in Tovi generic dataframe into REddyProc class}
\usage{
dfToREddyProcClass(df, siteName = "US-LICOR")
}
\arguments{
\item{df}{a dataframe}

\item{siteName}{a site name}
}
\value{
REddyProcClass
}
\description{
Conversion Function - converts data in Tovi generic dataframe into REddyProc class
}
